Commiting the mess...
It is a mess, but if I don't push this, nothing will happen.
driver.py can be run to start a server that you can connect to
using telnet. It attaches you to an object dynamically loaded
from the mudlib.

Actions don't work yet and dm-files are not used for this.
But the concept of the mudlib works.